refactor: move download related functions to separate module
+ renames and typehints
