+++ imaginairy/enhancers/describe_image_blip.py
@@ -9,7 +9,7 @@ from torchvision import transforms
 from torchvision.transforms.functional import InterpolationMode
 
 from imaginairy.utils import get_device
-from imaginairy.utils.model_manager import get_cached_url_path
+from imaginairy.utils.downloads import get_cached_url_path
 from imaginairy.vendored.blip.blip import BLIP_Decoder, load_checkpoint
 
 device = get_device()
