+++ imaginairy/img_processors/openpose.py
@@ -11,8 +11,8 @@ from scipy.ndimage.filters import gaussian_filter
 from torch import nn
 
 from imaginairy.utils import get_device
+from imaginairy.utils.downloads import get_cached_url_path
 from imaginairy.utils.img_utils import torch_image_to_openvcv_img
-from imaginairy.utils.model_manager import get_cached_url_path
 
 
 def pad_right_down_corner(img, stride, padValue):
