+++ scripts/controlnet_convert.py
@@ -3,7 +3,7 @@ import os
 import torch
 from safetensors.torch import load_file, save_file
 
-from imaginairy.utils.model_manager import get_cached_url_path
+from imaginairy.utils.downloads import get_cached_url_path
 from imaginairy.utils.paths import PKG_ROOT
 
 sd15_url = "https://huggingface.co/runwayml/stable-diffusion-v1-5/resolve/889b629140e71758e1e0006e355c331a5744b4bf/v1-5-pruned-emaonly.ckpt"
